revise infobar for tile_source

+++ imagedisplay.h
@@ -9,7 +9,7 @@ G_DECLARE_FINAL_TYPE( Imagedisplay, imagedisplay,
 void imagedisplay_image_to_gtk( Imagedisplay *imagedisplay, 
 	int x_image, int y_image, double *x_gtk, double *y_gtk );
 void imagedisplay_gtk_to_image( Imagedisplay *imagedisplay, 
-	double x_gtk, double y_gtk, int *x_image, int *y_image );
+	double x_gtk, double y_gtk, double *x_image, double *y_image );
 
 Imagedisplay *imagedisplay_new( TileCache *tile_cache );
